remove unused file; move library to a library
